drivers: Add fprotect driver (moved from bootloader)
So it can be used by others.
Also move tests and add rst docs.
