+++ drivers/CMakeLists.txt
@@ -13,4 +13,5 @@ add_subdirectory_ifdef(CONFIG_ADP536X adp536x)
 add_subdirectory_ifdef(CONFIG_ST25R3911B_LIB st25r3911b)
 add_subdirectory_ifdef(CONFIG_BT_LL_NRFXLIB bt_ll_nrfxlib)
 add_subdirectory_ifdef(CONFIG_NRF9160_GPS nrf9160_gps)
+add_subdirectory_ifdef(CONFIG_FPROTECT fprotect)
 add_subdirectory(flash_patch)
