+++ drivers/Kconfig
@@ -15,6 +15,7 @@ rsource "lte_link_control/Kconfig"
 rsource "net/Kconfig"
 rsource "sensor/Kconfig"
 rsource "st25r3911b/Kconfig"
+rsource "fprotect/Kconfig"
 rsource "flash_patch/Kconfig"
 
 if BT_LL_NRFXLIB
