+++ drivers/fprotect/CMakeLists.txt
