+++ drivers/fprotect/Kconfig
@@ -1,0 +1,12 @@
+#
+# Copyright (c) 2019 Nordic Semiconductor ASA
+#
+# SPDX-License-Identifier: LicenseRef-BSD-5-Clause-Nordic
+#
+
+config FPROTECT
+	bool "Write-protect flash regions"
+	depends on SOC_FAMILY_NRF
+	help
+	  Use hardware peripherals (BPROT, ACL, or SPU) to write-protect
+	  regions of flash until next reset.
