+++ drivers/fprotect/fprotect_spu.c
