+++ samples/bootloader/prj.conf
@@ -14,9 +14,10 @@ CONFIG_ERRNO=n
 CONFIG_SYS_CLOCK_EXISTS=n
 CONFIG_NRF_RTC_TIMER=n
 CONFIG_CLOCK_CONTROL=n
+CONFIG_FPROTECT=y
 CONFIG_SECURE_BOOT_DEBUG=y
 
-# To reduce size, turn off debug printing
+# To reduce size, turn off debug printing:
 # CONFIG_BOOT_BANNER=n
 # CONFIG_PRINTK=n
 # CONFIG_CONSOLE=n
