+++ tests/drivers/fprotect/negative/CMakeLists.txt
@@ -10,5 +10,3 @@ project(NONE)
 
 FILE(GLOB app_sources src/*.c)
 target_sources(app PRIVATE ${app_sources})
-target_include_directories(app PRIVATE .)
-target_include_directories(app PRIVATE ${ZEPHYR_BASE}/../nrf/subsys/bootloader/)
