+++ tests/drivers/fprotect/negative/prj.conf
@@ -9,7 +9,6 @@ CONFIG_ZTEST_STACKSIZE=8192
 CONFIG_TEST_USERSPACE=n
 CONFIG_USERSPACE=n
 CONFIG_STDOUT_CONSOLE=n
-CONFIG_FLOAT=y
 CONFIG_MPU_ALLOW_FLASH_WRITE=y
 CONFIG_SECURE_BOOT=y
 CONFIG_FLASH=y
