+++ tests/drivers/fprotect/positive/prj.conf
@@ -1,0 +1,14 @@
+#
+# Copyright (c) 2019 Nordic Semiconductor ASA
+#
+# SPDX-License-Identifier: LicenseRef-BSD-5-Clause-Nordic
+#
+
+CONFIG_ZTEST=y
+CONFIG_ZTEST_STACKSIZE=8192
+CONFIG_TEST_USERSPACE=n
+CONFIG_USERSPACE=n
+CONFIG_STDOUT_CONSOLE=n
+CONFIG_MPU_ALLOW_FLASH_WRITE=y
+CONFIG_SECURE_BOOT=y
+CONFIG_FLASH=y
